requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.7
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
scikit-learn==1.5.0
pymupdf==1.24.5
python-docx==1.1.2
streamlit==1.36.0
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.3.4
python-multipart==0.0.9
pytest==8.2.2
httpx==0.27.0
numpy==1.26.4
sentence-transformers==3.0.1
langchain-community==0.2.7
tiktoken==0.7.0

